+++ stripe/test/api_resources/abstract/test_api_resource.py
@@ -1,7 +1,9 @@
 import stripe
-from stripe.test.helper import (
-    StripeApiTestCase, MyResource, MyUpdateable, MySingleton
-)
+from stripe.test.helper import StripeApiTestCase
+
+
+class MyResource(stripe.api_resources.abstract.APIResource):
+    pass
 
 
 class APIResourceTests(StripeApiTestCase):
@@ -82,103 +84,3 @@ class APIResourceTests(StripeApiTestCase):
         for obj in [None, 1, 3.14, dict(), list(), set(), tuple(), object()]:
             self.assertRaises(stripe.error.InvalidRequestError,
                               MyResource.retrieve, obj)
-
-    def test_retrieve_and_update_with_stripe_version(self):
-        self.mock_response({
-            'id': 'foo',
-            'bobble': 'scrobble',
-        })
-
-        res = MyUpdateable.retrieve('foo', stripe_version='2017-08-15')
-
-        self.requestor_class_mock.assert_called_with(
-            account=None, api_base=None, key=None,
-            api_version='2017-08-15'
-        )
-
-        res.bobble = 'new_scrobble'
-        res.save()
-
-        self.requestor_class_mock.assert_called_with(
-            account=None, api_base=None, key='reskey',
-            api_version='2017-08-15'
-        )
-
-
-class SingletonAPIResourceTests(StripeApiTestCase):
-
-    def test_retrieve(self):
-        self.mock_response({
-            'single': 'ton'
-        })
-        res = MySingleton.retrieve()
-
-        self.requestor_mock.request.assert_called_with(
-            'get', '/v1/mysingleton', {}, None)
-
-        self.assertEqual('ton', res.single)
-
-
-class NestedResourceClassMethodsTests(StripeApiTestCase):
-    @stripe.api_resources.abstract.nested_resource_class_methods(
-        'nested',
-        operations=['create', 'retrieve', 'update', 'delete', 'list']
-    )
-    class MainResource(stripe.api_resources.abstract.APIResource):
-        pass
-
-    def test_create_nested(self):
-        self.mock_response({
-            'id': 'nested_id',
-            'object': 'nested',
-            'foo': 'bar',
-        })
-        nested_resource = self.MainResource.create_nested('id', foo='bar')
-        self.requestor_mock.request.assert_called_with(
-            'post', '/v1/mainresources/id/nesteds', {'foo': 'bar'}, None)
-        self.assertEqual('bar', nested_resource.foo)
-
-    def test_retrieve_nested(self):
-        self.mock_response({
-            'id': 'nested_id',
-            'object': 'nested',
-            'foo': 'bar',
-        })
-        nested_resource = self.MainResource.retrieve_nested('id', 'nested_id')
-        self.requestor_mock.request.assert_called_with(
-            'get', '/v1/mainresources/id/nesteds/nested_id', {}, None)
-        self.assertEqual('bar', nested_resource.foo)
-
-    def test_modify_nested(self):
-        self.mock_response({
-            'id': 'nested_id',
-            'object': 'nested',
-            'foo': 'baz',
-        })
-        nested_resource = self.MainResource.modify_nested('id', 'nested_id',
-                                                          foo='baz')
-        self.requestor_mock.request.assert_called_with(
-            'post', '/v1/mainresources/id/nesteds/nested_id', {'foo': 'baz'},
-            None)
-        self.assertEqual('baz', nested_resource.foo)
-
-    def test_delete_nested(self):
-        self.mock_response({
-            'id': 'nested_id',
-            'object': 'nested',
-            'deleted': True,
-        })
-        nested_resource = self.MainResource.delete_nested('id', 'nested_id')
-        self.requestor_mock.request.assert_called_with(
-            'delete', '/v1/mainresources/id/nesteds/nested_id', {}, None)
-        self.assertEqual(True, nested_resource.deleted)
-
-    def test_list_nesteds(self):
-        self.mock_response({
-            'object': 'list',
-            'data': [],
-        })
-        nested_resource = self.MainResource.list_nesteds('id')
-        self.requestor_mock.request.assert_called_with(
-            'get', '/v1/mainresources/id/nesteds', {}, None)
-        self.assertTrue(isinstance(nested_resource.data, list))
