+++ stripe/test/api_resources/abstract/test_singleton_api_resource.py
@@ -1,0 +1,20 @@
+import stripe
+from stripe.test.helper import StripeApiTestCase
+
+
+class MySingleton(stripe.api_resources.abstract.SingletonAPIResource):
+    pass
+
+
+class SingletonAPIResourceTests(StripeApiTestCase):
+
+    def test_retrieve(self):
+        self.mock_response({
+            'single': 'ton'
+        })
+        res = MySingleton.retrieve()
+
+        self.requestor_mock.request.assert_called_with(
+            'get', '/v1/mysingleton', {}, None)
+
+        self.assertEqual('ton', res.single)
