+++ stripe/test/api_resources/test_balance.py
@@ -13,21 +13,3 @@ class BalanceTest(StripeResourceTest):
             {},
             None
         )
-
-
-class BalanceTransactionTest(StripeResourceTest):
-
-    def test_list_balance_transactions(self):
-        stripe.BalanceTransaction.list()
-        self.requestor_mock.request.assert_called_with(
-            'get',
-            '/v1/balance/history',
-            {}
-        )
-
-    def test_convert_to_stripe_object(self):
-        transaction = stripe.util.convert_to_stripe_object({
-            'id': 'txn_foo',
-            'object': 'balance_transaction',
-        })
-        self.assertIsInstance(transaction, stripe.BalanceTransaction)
