+++ stripe/test/helper.py
@@ -17,12 +17,6 @@ DUMMY_CHARGE = {
     'source': 'tok_visa'
 }
 
-DUMMY_DISPUTE = {
-    'status': 'needs_response',
-    'currency': 'usd',
-    'metadata': {}
-}
-
 DUMMY_PLAN = {
     'amount': 2000,
     'interval': 'month',
@@ -31,88 +25,6 @@ DUMMY_PLAN = {
     'id': ('stripe-test-gold-' +
            ''.join(random.choice(string.ascii_lowercase) for x in range(10)))
 }
-
-DUMMY_COUPON = {
-    'percent_off': 25,
-    'duration': 'repeating',
-    'duration_in_months': 5,
-    'metadata': {}
-}
-
-DUMMY_RECIPIENT = {
-    'name': 'John Doe',
-    'type': 'individual'
-}
-
-DUMMY_TRANSFER = {
-    'amount': 400,
-    'currency': 'usd',
-    'recipient': 'self'
-}
-
-DUMMY_APPLE_PAY_DOMAIN = {
-    'domain_name': 'test.com',
-}
-
-DUMMY_INVOICE_ITEM = {
-    'amount': 456,
-    'currency': 'usd',
-}
-
-SAMPLE_INVOICE = stripe.util.json.loads("""
-{
-  "amount_due": 1305,
-  "attempt_count": 0,
-  "attempted": true,
-  "charge": "ch_wajkQ5aDTzFs5v",
-  "closed": true,
-  "customer": "cus_osllUe2f1BzrRT",
-  "date": 1338238728,
-  "discount": null,
-  "ending_balance": 0,
-  "id": "in_t9mHb2hpK7mml1",
-  "livemode": false,
-  "next_payment_attempt": null,
-  "object": "invoice",
-  "paid": true,
-  "period_end": 1338238728,
-  "period_start": 1338238716,
-  "starting_balance": -8695,
-  "subtotal": 10000,
-  "total": 10000,
-  "lines": {
-    "invoiceitems": [],
-    "prorations": [],
-    "subscriptions": [
-      {
-        "plan": {
-          "interval": "month",
-          "object": "plan",
-          "identifier": "expensive",
-          "currency": "usd",
-          "livemode": false,
-          "amount": 10000,
-          "name": "Expensive Plan",
-          "trial_period_days": null,
-          "id": "expensive"
-        },
-        "period": {
-          "end": 1340917128,
-          "start": 1338238728
-        },
-        "amount": 10000
-      }
-    ]
-  }
-}
-""")
-
-DUMMY_WEBHOOK_PAYLOAD = """{
-  "id": "evt_test_webhook",
-  "object": "event"
-}"""
-
-DUMMY_WEBHOOK_SECRET = 'whsec_test_secret'
 
 
 class StripeTestCase(unittest2.TestCase):
@@ -190,34 +102,3 @@ class StripeResourceTest(StripeApiTestCase):
     def setUp(self):
         super(StripeResourceTest, self).setUp()
         self.mock_response({})
-
-
-class MyResource(stripe.api_resources.abstract.APIResource):
-    pass
-
-
-class MySingleton(stripe.api_resources.abstract.SingletonAPIResource):
-    pass
-
-
-class MyListable(stripe.api_resources.abstract.ListableAPIResource):
-    pass
-
-
-class MyCreatable(stripe.api_resources.abstract.CreateableAPIResource):
-    pass
-
-
-class MyUpdateable(stripe.api_resources.abstract.UpdateableAPIResource):
-    pass
-
-
-class MyDeletable(stripe.api_resources.abstract.DeletableAPIResource):
-    pass
-
-
-class MyComposite(stripe.api_resources.abstract.ListableAPIResource,
-                  stripe.api_resources.abstract.CreateableAPIResource,
-                  stripe.api_resources.abstract.UpdateableAPIResource,
-                  stripe.api_resources.abstract.DeletableAPIResource):
-    pass
