+++ stripe/test/test_stripe_object.py
@@ -4,7 +4,56 @@ from copy import copy, deepcopy
 
 import stripe
 from stripe import util
-from stripe.test.helper import StripeUnitTestCase, SAMPLE_INVOICE
+from stripe.test.helper import StripeUnitTestCase
+
+
+SAMPLE_INVOICE = stripe.util.json.loads("""
+{
+  "amount_due": 1305,
+  "attempt_count": 0,
+  "attempted": true,
+  "charge": "ch_wajkQ5aDTzFs5v",
+  "closed": true,
+  "customer": "cus_osllUe2f1BzrRT",
+  "date": 1338238728,
+  "discount": null,
+  "ending_balance": 0,
+  "id": "in_t9mHb2hpK7mml1",
+  "livemode": false,
+  "next_payment_attempt": null,
+  "object": "invoice",
+  "paid": true,
+  "period_end": 1338238728,
+  "period_start": 1338238716,
+  "starting_balance": -8695,
+  "subtotal": 10000,
+  "total": 10000,
+  "lines": {
+    "invoiceitems": [],
+    "prorations": [],
+    "subscriptions": [
+      {
+        "plan": {
+          "interval": "month",
+          "object": "plan",
+          "identifier": "expensive",
+          "currency": "usd",
+          "livemode": false,
+          "amount": 10000,
+          "name": "Expensive Plan",
+          "trial_period_days": null,
+          "id": "expensive"
+        },
+        "period": {
+          "end": 1340917128,
+          "start": 1338238728
+        },
+        "amount": 10000
+      }
+    ]
+  }
+}
+""")
 
 
 class StripeObjectTests(StripeUnitTestCase):
